move scripts to a separate folder for better org
