--- script-opts/detect_image.conf
@@ -1,17 +1,0 @@
-# commands to execute when a file detected as an image (1 frame, no audio) is loaded or unloaded
-
-# an image was loaded, and the previous file was not an image (or there was no previous file)
-command_on_first_image_loaded=
-# an image was loaded (regardless of what the previous file was)
-command_on_image_loaded=
-# a non-image was loaded, and the previous file was an image
-command_on_non_image_loaded=
-
-# the purpose of these "hooks" is to let you change bindings, profiles, reset properties...
-# see https://mpv.io/manual/master/#list-of-input-commands for general command information
-# note that there is no such thing as "unloading a profile", to emulate this you must create an opposite profile and load that
-
-# example possible values:
-#command_on_first_image_loaded=apply-profile image; enable-section image-viewer; script-message status-line-enable
-#command_on_image_loaded=no-osd set video-pan-x 0; script-message align-border "" -1
-#command_on_non_image_loaded=disable-section image-viewer; no-osd set video-pan-x 0; no-osd set video-pan-y 0; no-osd set video-zoom 0; script-message status-line-disable
