--- script-opts/image_positioning.conf
@@ -1,14 +1,0 @@
-# size of the margins with drag-to-pan
-drag_to_pan_margin=50
-drag_to_pan_move_if_full_view=no
-
-# size of the margins with pan-follows-cursor
-pan_follows_cursor_margin=50
-
-# size of the margins with cursor-centric-zoom
-cursor_centric_zoom_margin=50
-# if the borders would show up, move the image 
-# this makes it not exactly cursor-centric in some cases
-cursor_centric_zoom_auto_center=yes
-# allow zooming out if the image can already fully fit in the window
-cursor_centric_zoom_dezoom_if_full_view=no
