--- script-opts/minimap.conf
@@ -1,19 +1,0 @@
-# whether to show by default
-enabled=yes
-# the position of the center of the minimap, in percentage of the window (x, y)
-center=92,92
-# the scale of the minimap (i.e. the view rectangle is scale / 100 times the size of the window)
-scale=12
-# the cutoff size of the minimap (i.e. the image rectangle is clipped if it falls outside of the this zone)
-max_size=16,16
-# opacity of the "image" (from 00=opaque to FF=transparent)
-image_opacity=88
-# color of the "image" (#BBGGRR where each component rages from 00 to FF)
-image_color=BBBBBB
-# opacity of the "view"
-view_opacity=BB
-view_color=222222
-# whether the view should be drawn above the image
-view_above_image=yes
-# whether to show the minimap if the current image is fully visible
-hide_when_full_image_in_view=yes
