+++ script-opts/mvi/1_detect_image.conf
@@ -1,0 +1,11 @@
+# MUST be 'key= value', no spaces before '=' or not loaded!!!
+# commands to execute when a file detected as an image (1 frame, no audio) is loaded or unloaded
+# these "hooks" let you change bindings, profiles, reset properties... mpv.io/manual/master/#list-of-input-commands for general command information
+# CAN'T "unload a profile", to emulate this you must create an opposite profile and load that
+
+#on_load_image_first	= loaded     image, and the previous file was not an image (or there was no previous file)
+#on_load_image      	= loaded     image (regardless of what the previous file was)
+#on_load_non_image  	= loaded non-image, and the previous file was an image
+on_load_image_first=	apply-profile mvi        ; enable-section  mvi;                                                                              script-message status-line-enable
+on_load_image=      	                                                                          no-osd set video-pan-x 0;                          script-message align-border "" -1
+on_load_non_image=  	apply-profile mvi restore; disable-section mvi; no-osd set video-pan-y 0; no-osd set video-pan-x 0; no-osd set video-zoom 0; script-message status-line-disable
