+++ script-opts/mvi/3_status_line.conf
@@ -1,0 +1,12 @@
+enabled=yes
+#↑ whether to show by default
+size=36
+#↑ its font size
+margin=10
+#↑ distance of the text to the borders
+text_top_left=
+text_top_right=
+text_bottom_left=${filename} [${playlist-pos-1}/${playlist-count}]
+text_bottom_right=[${dwidth:X}x${dheight:X}]
+# text to be expanded, see property expansion: https://mpv.io/manual/master/#property-expansion
+# \N line breaks, can also use ass tags, see here: http://docs.aegisub.org/3.2/ASS_Tags/
