+++ script-opts/mvi/4_minimap.conf
@@ -1,0 +1,22 @@
+# boolean/numbers can't be converted with comments by github.com/mpv-player/mpv/blob/master/player/lua/options.lua
+# other values also bug due to script imports not expecting comments
+enabled=yes
+#↑ whether to show by default
+scale=12
+#↑ the scale of the minimap (i.e. the view rectangle is scale / 100 times the size of the window)
+view_above_image=yes
+#↑ whether the view should be drawn above the image
+hide_when_full_image_in_view=yes
+#↑ whether to show the minimap if the current image is fully visible
+center=92,92
+#↑ the position of the center of the minimap, in percentage of the window (x, y)
+max_size=16,16
+#↑ the cutoff size of the minimap (i.e. the image rectangle is clipped if it falls outside of the this zone)
+image_opacity=88
+#↑ opacity of the "image" (from 00=opaque to FF=transparent)
+image_color=BBBBBB
+#↑ color of the "image" (#BBGGRR where each component rages from 00 to FF)
+view_opacity=BB
+#↑ opacity of the "view"
+view_color=222222
+#↑
