+++ script-opts/mvi/5_ruler.conf
@@ -1,0 +1,21 @@
+show_distance=yes
+#↑ whether to show the length of the lines between the two points
+show_coordinates=yes
+#↑ whether to show the coordinates of the two points
+coordinates_space=image
+#↑ the coordinate space of the text shown. Can be "image", "window", "both"
+show_angles=degrees
+#↑ can be "degrees", "radians", "both", or "no"
+line_width=2
+dots_radius=3
+font_size=36
+line_color=33
+#↑ ranges from 00 (black) to FF (white)
+confirm_bindings=MBTN_LEFT,ENTER
+#↑ bindings used to set points. The binding to trigger ruler mode can also be used. Comma-separated list
+exit_bindings=ESC
+#↑ bindings used to set points. The binding to trigger ruler mode can also be used. Comma-separated list
+set_first_point_on_begin=no
+#↑ if yes, the first point will be immediately set at the cursor position when calling 'ruler'
+clear_on_second_point_set=no
+#↑ if yes, the ruler overlay will be immediately cleared when setting the second point
