--- script-opts/ruler.conf
@@ -1,21 +1,0 @@
-# whether to show the length of the lines between the two points
-show_distance=yes
-# whether to show the coordinates of the two points
-show_coordinates=yes
-# the coordinate space of the text shown. Can be "image", "window", "both"
-coordinates_space=image
-# can be "degrees", "radians", "both", or "no"
-show_angles=degrees
-line_width=2
-dots_radius=3
-font_size=36
-# ranges from 00 (black) to FF (white)
-line_color=33
-# bindings used to set points. The binding to trigger ruler mode can also be used. Comma-separated list
-confirm_bindings=MBTN_LEFT,ENTER
-# bindings used to set points. The binding to trigger ruler mode can also be used. Comma-separated list
-exit_bindings=ESC
-# if yes, the first point will be immediately set at the cursor position when calling 'ruler'
-set_first_point_on_begin=no
-# if yes, the ruler overlay will be immediately cleared when setting the second point
-clear_on_second_point_set=no
