--- script-opts/status_line.conf
@@ -1,14 +1,0 @@
-# whether to show by default
-enabled=yes
-# its font size
-size=36
-# distance of the text to the borders
-margin=10
-# the text to be expanded
-# see property expansion: https://mpv.io/manual/master/#property-expansion
-# \N can be used for line breaks
-# you can also use ass tags, see here: http://docs.aegisub.org/3.2/ASS_Tags/
-text_top_left=
-text_top_right=
-text_bottom_left=${filename} [${playlist-pos-1}/${playlist-count}]
-text_bottom_right=[${dwidth:X}x${dheight:X}]
